+++ Source/Player/Camera/VRCustomCamera.cs
@@ -1,6 +1,7 @@
 using System;
 using RepoXR.Input;
 using UnityEngine;
+using UnityEngine.Experimental.Rendering;
 using UnityEngine.UI;
 
 namespace RepoXR.Player.Camera;
@@ -12,40 +13,49 @@ namespace RepoXR.Player.Camera;
 public class VRCustomCamera : MonoBehaviour
 {
     public static VRCustomCamera instance;
-    
+
+    private RenderTexture targetTexture;
+
     [SerializeField] protected UnityEngine.Camera mainCamera;
     [SerializeField] protected UnityEngine.Camera topCamera;
     [SerializeField] protected UnityEngine.Camera uiCamera;
 
     [SerializeField] protected Image overlayImage;
-    
+
     private Transform gameplayCamera;
+
+    private int lastWidth;
+    private int lastHeight;
+
+    private float frameTimer;
 
     private void Awake()
     {
         instance = this;
-        
+
         var fov = Plugin.Config.CustomCameraFOV.Value;
 
         mainCamera.fieldOfView = fov;
         topCamera.fieldOfView = fov;
         uiCamera.fieldOfView = fov;
-        
+
         gameplayCamera = UnityEngine.Camera.main!.transform;
-        
+
         transform.localPosition = TrackingInput.Instance.HeadTransform.localPosition;
         transform.localRotation = TrackingInput.Instance.HeadTransform.localRotation;
-        
+
         Plugin.Config.CustomCameraFOV.SettingChanged += OnFOVChanged;
+
+        UpdateRenderTexture();
     }
 
     private void OnDestroy()
     {
         instance = null!;
-        
+
         Plugin.Config.CustomCameraFOV.SettingChanged -= OnFOVChanged;
     }
-    
+
     private void OnFOVChanged(object sender, EventArgs e)
     {
         var fov = Plugin.Config.CustomCameraFOV.Value;
@@ -69,14 +79,58 @@ public class VRCustomCamera : MonoBehaviour
         // Some weird fog thing, I don't know why but this is needed
         RenderSettings.fogDensity =
             SemiFunc.MenuLevel() || SemiFunc.RunIsShop() || SemiFunc.RunIsLobby() ? 0.015f : 0.15f;
+
+        if (lastWidth != Screen.width || lastHeight != Screen.height)
+            UpdateRenderTexture();
+
+        frameTimer += Time.unscaledDeltaTime;
+
+        var interval = 1f / Plugin.Config.CustomCameraFramerate.Value;
+        if (frameTimer < interval)
+            return;
+
+        frameTimer -= interval;
+
+        mainCamera.Render();
+        topCamera.Render();
+        uiCamera.Render();
     }
 
-    
     private void LateUpdate()
     {
         // Since we override the FadeOverlay image color in a LateUpdate, we need to read it back in a late update as well
         // Also this script needs to execute *after* the override, hence the [DefaultExecutionOrder(100)]
         overlayImage.color = FadeOverlay.Instance.Image.color;
+    }
+
+    private void UpdateRenderTexture()
+    {
+        lastWidth = Screen.width;
+        lastHeight = Screen.height;
+
+        if (targetTexture != null)
+            targetTexture.Release();
+
+        targetTexture = new RenderTexture(lastWidth, lastHeight, GraphicsFormat.R8G8B8A8_UNorm,
+            GraphicsFormat.D32_SFloat_S8_UInt)
+        {
+            name = "Custom Camera RT",
+            antiAliasing = 1,
+            useMipMap = false,
+            autoGenerateMips = false,
+        };
+
+        mainCamera.targetTexture = targetTexture;
+        topCamera.targetTexture = targetTexture;
+        uiCamera.targetTexture = targetTexture;
+    }
+
+    private void OnRenderImage(RenderTexture source, RenderTexture destination)
+    {
+        if (targetTexture != null)
+            Graphics.Blit(targetTexture, destination);
+        else
+            Graphics.Blit(Texture2D.blackTexture, destination);
     }
 }